Refactor VR case report backend exporters and fixe some bugs
- Shared functionality in interface, and common endpoint for all
- Use streams directly in place of byte arrays
- Needs filtering and sorting to be checked
- The "status" check is maybe not perfect, but it's the same as in the
  frontend
